use flameshot on x11

+++ screenshot.zsh
@@ -4,10 +4,16 @@ set -e
 
 # screenshot
 
-# flameshot gui
+if [[ $XDG_SESSION_TYPE = 'wayland' ]]; then
 
-FILE=/tmp/screenshot.png
-gnome-screenshot --area --file $FILE
-satty --filename $FILE
-rm $FILE
+  FILE=/tmp/screenshot.png
+  gnome-screenshot --area --file $FILE
+  satty --filename $FILE
+  rm $FILE
+
+else
+
+  flameshot gui
+
+fi
